+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 + Added **Visitor Display** - Show all items needed for the visitors.
 + Added **Highlight Ready** - Highlight the visitor when the required items are in the inventory.
 + Added **Show Price** - Show the bazaar price of the items required for the visitors.
++ Added **Crop Milestone** Number - Show the number of the crop milestone in the inventory.
++ Added **Crop Upgrades** Number - Show the number of upgrades in the crop upgrades inventory.
 
 ### Features from other Mods
 > *The following features are only there because I want them when testing SkyHanni features without other mods present.*